django/requirements.txt
Django==5.2.6
djangorestframework==3.15.0
psycopg2-binary==2.9.9
requests==2.31.0
python-decouple==3.8
django-cors-headers==4.4.0